+++ CHANGELOG.md
@@ -24,7 +24,7 @@ to [sourceforge feature requests](https://sourceforge.net/p/jabref/features/) by
 - Fixed #417: Table now updates when switching groups
 
 ### Removed
-
+- OpenOffice/LibreOffice is removed from the push-to-application button and only accessed through the side panel